requirements.txt
aiohttp==3.6.2
aiomysql==0.0.20
altgraph==0.17
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
discord.py==1.3.4
dnspython==1.16.0
docopt==0.6.2
ffmpeg==1.4
future==0.18.2
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-images-download==2.8.0
googleapis-common-protos==1.51.0
googletrans==2.4.0
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
itsdangerous==1.1.0
MarkupSafe==1.1.1
multidict==4.7.5
murmurhash==1.0.2
opuslib==3.0.1
pefile==2019.4.18
Pillow==7.1.1
pipreqs==0.4.10
praw==6.5.1
prawcore==1.0.1
preshed==3.0.2
protobuf==3.11.3
pycparser==2.19
PyDrive==1.3.1
PyMySQL==0.9.2
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.23.0
six==1.14.0
update-checker==0.16
urllib3==1.25.8
websocket-client==0.57.0
websockets==8.1
Werkzeug==0.16.0
yarg==0.1.9
yarl==1.4.2
youtube-dl==2020.3.8
